requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
demjson==2.2.4
dnspython==2.3.0
exceptiongroup==1.1.1
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
idna==3.4
jtutils==0.0.8
kiwisolver==1.4.4
leven==1.0.4
matplotlib==3.7.1
multidict==6.0.4
nose==1.3.7
numpy==1.24.3
openai==0.27.7
outcome==1.2.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-csv==0.0.13
python-dateutil==2.8.2
python-whois==0.8.0
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tld==0.13
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==2.0.2
wsproto==1.2.0
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.9.2
